refactor(api-settings): split settings page components
